+++ R/st_nn_poly.R
@@ -1,4 +1,4 @@
-.st_nn_poly = function(x, y, k, maxdist) {
+.st_nn_poly = function(x, y, k, maxdist, progress) {
 
   x_features = length(sf::st_geometry(x))
   maxdist = units::set_units(maxdist, "m")
@@ -7,11 +7,13 @@
   dist_matrix = matrix(NA, nrow = x_features, ncol = k)
 
   # Progress bar
-  pb = utils::txtProgressBar(min = 0, max = x_features, initial = 0, style = 3)
+  if(progress) {
+    pb = utils::txtProgressBar(min = 0, max = x_features, initial = 0, style = 3)
+  }
 
   for(i in 1:x_features) {
 
-    dists = sf::st_distance(x[i, ], y)[1, ]
+    dists = sf::st_distance(x[i], y)[1, ]
     ids1 = order(dists)[1:k]
     dists1 = dists[ids1]
     dists1 = units::set_units(dists1, "m")
@@ -21,11 +23,15 @@
     dist_matrix[i, ] = dists1
 
     # Progress
-    utils::setTxtProgressBar(pb, i)
+    if(progress) {
+      utils::setTxtProgressBar(pb, i)
+    }
 
   }
 
-  cat("\n")
+  if(progress) {
+    cat("\n")
+  }
 
   return(list(ids, dist_matrix))